I improving integral performance.
I reworked the integral loops to try and remove as much as possible to
outer loops. Also I hacked on TA a little to get rid of one of the
memory allocations inside the inner loop. This code still needs cleaning
up and testing, but both memory pressure and performance seem better
than with old integrals.